+++ packages/pangea-sdk/tests/integration2/test_authn.py
@@ -28,7 +28,8 @@ def client(request: pytest.FixtureRequest) -> Iterator[AuthN]:
 
 @pytest.fixture(scope="session")
 async def async_client(request: pytest.FixtureRequest) -> AsyncIterator[AuthNAsync]:
-    yield AuthNAsync(token="my_api_token", config=PangeaConfig(base_url_template=base_url))
+    async with AuthNAsync(token="my_api_token", config=PangeaConfig(base_url_template=base_url)) as client:
+        yield client
 
 
 class TestAuthN: