+++ packages/pangea-sdk/CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Made exports more explicit.
+- Retries now include a request header `X-Pangea-Retried-Request-Ids` to track
+  the request IDs of the retries.
 
 ### Fixed